requirements.txt
openai
python-docx
pypdf
streamlit==1.37.1
tiktoken
